bot/pyproject.toml
[tool.poetry]
name = "discord-mediawiki-bot"
version = "0.1.0"
description = "Discord Bot for MediaWiki Verification"
authors = ["Your Name"]
readme = "README.md"
packages = [{ include = "bot" }]

[tool.poetry.dependencies]
python = "^3.12"
discord.py = "^2.3.2"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"